A few small adjustments
You were using the same image instead of the random when you were creating the new card on line 48.
Additionally, line 74 there was a typo on the right side of the equality check.
And finally, the setTimeout was being canceled immediate  (it was never called) because the clear Timeout was outside of the timeout function. With it inside the timeout function, the loop will start (line 86) and then the timeout is cleared so it can be used again for the next pair of guesses.
So you were super close. Just a few small tweaks!
Good job so far!